RavenDB-18085 introduced StudioCollectionsHandlerProcessorForPreviewCollection
